qwen3以及vllm适配
* 针对jnu_server进行修改

* 更新env文件

* 更新env文件

* 更新protocal文件

* 修改了adapter写成apapter的错误

* 删除了部分导致报错的多余模块

* 修复api/core的部分错误

* 修复api/core的部分错误

* 修复了部分bug

* 修复了部分bug

* 进行qwen3思考模式的适配

* qwen3支持开发

* qwen3支持开发

* qwen3支持开发

* qwen3支持开发

* qwen3支持开发

* qwen3支持开发

* qwen3支持开发

* 使得topK参数可使用

* 开发进度保存

* 开发vllm适配

* 开发vllm适配

* 开发vllm适配

* 开发vllm适配

* 开发vllm适配

* 开发vllm适配

* 开发vllm适配

* 开发vllm适配

* 开发vllm适配

* qwen3适配

+++ api/utils/protocol.py
@@ -158,7 +158,7 @@ class ChatCompletionCreateParams(BaseModel):
     We generally recommend altering this or `temperature` but not both.
     """
 
-    top_k: Optional[float] = 20
+    top_k: Optional[int] = 20
     """
     生成过程中采样候选集的大小。
     例如,取值为50时,仅将单次生成中得分最高的50个Token组成随机采样的候选集。
